multiplas alteracoes

+++ run.sh
@@ -1,8 +1,17 @@
 #!/bin/bash
 
-# Cria Log de tudo
-LOGFILE="~/vps_config.log"
+#INICIO DO GERA LOG DE TUDO
+LOGFILE="/root/vps_config.log"
 exec > >(tee -a $LOGFILE) 2>&1
+#FIM DO GERA LOG DE TUDO ################################################
+
+#VALIDA VERSAO DO LINUX
+LINUXRELEASE=$(lsb_release -d | awk '{print $2,$3}')
+if [[ $LINUXRELEASE =~ "Ubuntu 23." ]]; then
+    PIPCOMMAND="pip install --break-system-packages"
+else
+    PIPCOMMAND="pip install"
+fi
 
 #INICIO DE INSTALACAO DE PRE-REQUISITOS #################################
 
@@ -22,6 +31,7 @@ setupEnvironment() {
     LOCALPATH=$(pwd)
     TOOLSPATH="/opt/tools"
     WLPATH="/opt/wordlists"
+    DIRBINPATH="/usr/local/bin"
 
     #Diretorio de ferramentas
     mkdir /opt/tools
@@ -75,21 +85,38 @@ setupGolang () {
     #https://miek.nl/2020/july/17/script-to-upgrade-to-latest-go-version/
     local LATEST=$(curl -s 'https://go.dev/dl/?mode=json' | jq -r '.[0].version')
     local INSTALLED=$(go version | awk '{ print $3 }')
-    if [[ ${INSTALLED} == ${LATEST} ]]; then
-        echo Go is up to date, running ${LATEST} >&2
-        exit 0
-    fi
-    echo Upgrading Go from ${INSTALLED} to ${LATEST} >&2
-    local GOLANG=https://dl.google.com/go/${LATEST}.linux-amd64.tar.gz
-    local TAR=$(basename $GOLANG)
+    if ! command -v go ; then
+        echo -e "${RED}Installing Go version ${LATEST}"
+        local GOLANG=https://dl.google.com/go/${LATEST}.linux-amd64.tar.gz
+        local TAR=$(basename $GOLANG)
+        (   cd ${GODIR}
+            echo Downloading and extracting: $GOLANG
+            wget -q $GOLANG && tar xvfz ${TAR}
+        )
+    else
+        if [[ ${INSTALLED} == ${LATEST} ]]; then
+            echo Go is up to date, running ${LATEST} >&2
+            exit 0
+        fi
+        echo Upgrading Go from ${INSTALLED} to ${LATEST} >&2
+        local GOLANG=https://dl.google.com/go/${LATEST}.linux-amd64.tar.gz
+        local TAR=$(basename $GOLANG)
 
-    ( cd ${GODIR}
-      echo Downloading and extracting: $GOLANG >&2
-      wget -q $GOLANG && rm -rf go && tar xvfz ${TAR}
-    )
+        ( cd ${GODIR}
+          echo Downloading and extracting: $GOLANG >&2
+          wget -q $GOLANG && rm -rf go && tar xvfz ${TAR}
+        )
+    fi
 
     export PATH=$PATH:/usr/local/go/bin
+    export GOPATH=/root/go
+    export GOCACHE=/root/go/cache
     go version
+
+    # Adiciona ao .bashrc o export
+    echo 'export PATH=$PATH:/usr/local/go/bin' >> /root/.bashrc
+    echo 'export GOPATH=/root/go' >> /root/.bashrc
+    echo 'export GOCACHE=/root/go/cache' >> /root/.bashrc
 }
 #FIM DE INSTALACAO DE PRE-REQUISITOS #################################
 
@@ -97,7 +124,7 @@ setupGolang () {
 
 Altdns(){
     echo -e "${RED}[+]${FUNCNAME[0]}${NC}"
-    pip3 install py-altdns==1.0.2
+    $PIPCOMMAND py-altdns==1.0.2 
 }
 
 Alterx(){
@@ -120,14 +147,14 @@ Antiburl(){
     cd ${TOOLSPATH} 
     wget https://raw.githubusercontent.com/tomnomnom/hacks/master/anti-burl/main.go
     go build main.go
-    rm -rf main.go
-    mv main anti-burl ; chmod +x anti-burl
-    ln -s $TOOLSPATH/anti-burl /usr/local/bin/anti-burl
+    rm -f main.go
+    mv main anti-burl && chmod +x anti-burl
+    ln -s ${TOOLSPATH}/anti-burl ${DIRBINPATH}/anti-burl
 }
 
 Arjun(){
     echo -e "${RED}[+]${FUNCNAME[0]}${NC}"
-    pip3 install arjun
+    $PIPCOMMAND arjun
 }
 
 Assetfinder(){
@@ -142,7 +169,7 @@ Burl(){
 
 ChaosClient(){
     echo -e "${RED}[+]${FUNCNAME[0]}${NC}"
-    GO111MODULE=on go get -v github.com/projectdiscovery/chaos-client/cmd/chaos
+    go install -v github.com/projectdiscovery/chaos-client/cmd/chaos@latest
 }
 
 Collector(){
@@ -151,7 +178,7 @@ Collector(){
     wget https://raw.githubusercontent.com/m4ll0k/Bug-Bounty-Toolz/master/collector.py
     sed -i '#!/usr/bin/env python3' collector.py
     chmod +x collector.py
-    ln -s ${TOOLSPATH}/collector.py /usr/local/bin/collector.py
+    ln -s ${TOOLSPATH}/collector.py ${DIRBINPATH}/collector.py
 }
 
 DalFox(){
@@ -161,17 +188,12 @@ DalFox(){
 
 Dirsearch(){
     echo -e "${RED}[+]${FUNCNAME[0]}${NC}"
-    cd ${TOOLSPATH}
-    git clone https://github.com/maurosoria/dirsearch.git
-    cd dirsearch
-    pip3 install -r requirements.txt
-    chmod +x dirsearch.py
-    ln -s ${TOOLSPATH}/dirsearch/dirsearch.py /usr/local/bin/dirsearch
+    $PIPCOMMAND dirsearch
 }
 
 Dnsgen(){
     echo -e "${RED}[+]${FUNCNAME[0]}${NC}"
-    pip3 install dnsgen
+    $PIPCOMMAND dnsgen
 }
 
 Dnsx(){
@@ -197,8 +219,8 @@ Findomains(){
     cd /tmp
     wget https://github.com/findomain/findomain/releases/latest/download/findomain-linux.zip
     unzip findomain-linux.zip
-    mv findomain /usr/local/bin
-    chmod +x /usr/local/bin/findomain
+    mv findomain ${DIRBINPATH}
+    chmod +x ${DIRBINPATH}/findomain
 }
 
 Gau(){
@@ -219,7 +241,8 @@ Gf(){
 
     mkdir ~/.gf
     cp -r examples/* ~/.gf
-    cd /tmp ; git clone https://github.com/1ndianl33t/Gf-Patterns ; cd Gf-Patterns ; cp *.json ~/.gf
+    cd /tmp
+    git clone https://github.com/1ndianl33t/Gf-Patterns && cd Gf-Patterns && cp *.json ~/.gf
 }
 
 GithubSearch(){
@@ -227,8 +250,8 @@ GithubSearch(){
     cd ${TOOLSPATH}
     git clone https://github.com/gwen001/github-search.git
     cd github-search
-    pip3 install -r requirements3.txt
-    ln -s ${TOOLSPATH}/github-search/github-subdomains.py /usr/local/bin/github-subdomains
+    $PIPCOMMAND -r requirements.txt
+    ln -s ${TOOLSPATH}/github-search/github-subdomains.py ${DIRBINPATH}/github-subdomains
 }
 
 
@@ -237,14 +260,14 @@ GitDorker(){
     cd ${TOOLSPATH}
     git clone https://github.com/obheda12/GitDorker.git
     cd GitDorker
-    pip3 install -r requirements.txt
-    chmod +x chmod +x GitDorker.py
-    ln -s ${TOOLSPATH}/GitDorker/GitDorker.py /usr/local/bin/gitdorker
+    $PIPCOMMAND -r requirements.txt
+    chmod +x GitDorker.py
+    ln -s ${TOOLSPATH}/GitDorker/GitDorker.py ${DIRBINPATH}/gitdorker
 }
 
 GitDumper(){
     echo -e "${RED}[+]${FUNCNAME[0]}${NC}"
-    pip install git-dumper
+    $PIPCOMMAND git-dumper
 }
 
 GoogleChrome(){
@@ -261,10 +284,7 @@ GoSpider(){
 
 Gowitness(){
     echo -e "${RED}[+]${FUNCNAME[0]}${NC}"
-    local GOWITNESS=https://github.com/sensepost/gowitness/releases/download/2.4.0/gowitness-2.4.0-linux-amd64
-    cd $TOOLSPATH ; wget $GOWITNESS
-    chmod +x $(basename $GOWITNESS)
-    ln -s $TOOLSPATH/$(basename $GOWITNESS) /usr/local/bin/gowitness
+    go install -v github.com/sensepost/gowitness@latest
 }
 
 Hakrawler(){
@@ -298,9 +318,9 @@ JSScanner(){
     cd ${TOOLSPATH}
     git clone https://github.com/0x240x23elu/JSScanner.git
     cd JSScanner
-    pip3 install -r requirements.txt
+    $PIPCOMMAND -r requirements.txt
     chmod +x JSScanner.py
-    ln -s ${TOOLSPATH}/JSScanner/JSScanner.py /usr/local/bin/jsscanner
+    ln -s ${TOOLSPATH}/JSScanner/JSScanner.py ${DIRBINPATH}/jsscanner
 }
 
 JsubFinder(){
@@ -315,7 +335,7 @@ Kiterunner(){
     wget https://github.com/assetnote/kiterunner/releases/download/v1.0.2/kiterunner_1.0.2_linux_amd64.tar.gz
     tar xvzf kiterunner_1.0.2_linux_amd64.tar.gz
     chmod +x kr 
-    ln -s ${TOOLSPATH}/kr /usr/local/bin/kr
+    ln -s ${TOOLSPATH}/kr ${DIRBINPATH}/kr
 }
 
 LinkFinder(){
@@ -323,9 +343,9 @@ LinkFinder(){
     cd ${TOOLSPATH}
     git clone https://github.com/GerbenJavado/LinkFinder
     cd LinkFinder
-    pip3 install -r requirements.txt
+    $PIPCOMMAND -r requirements.txt
     python3 setup.py install
-    ln -s ${TOOLSPATH}/LinkFinder/linkfinder.py /usr/local/bin/linkfinder
+    ln -s ${TOOLSPATH}/LinkFinder/linkfinder.py ${DIRBINPATH}/linkfinder
 }
 
 Mapcidr(){
@@ -339,12 +359,12 @@ Massdns(){
     git clone https://github.com/blechschmidt/massdns.git
     cd massdns
     make && make nolinux
-    ln -s $TOOLSPATH/massdns/bin/massdns /usr/local/bin/massdns
+    cp -v ${TOOLSPATH}/massdns/bin/massdns ${DIRBINPATH}/massdns
 }
 
 MegaPy(){
     echo -e "${RED}[+]${FUNCNAME[0]}${NC}"
-    pip install mega.py
+    $PIPCOMMAND mega.py
 }
 
 Metabigor(){
@@ -388,8 +408,8 @@ ParamSpider(){
     cd ${TOOLSPATH}
     git clone https://github.com/devanshbatham/ParamSpider
     cd ParamSpider
-    pip install .
-    ln -s ${TOOLSPATH}/ParamSpider/paramspider.py /usr/local/bin/paramspider
+    $PIPCOMMAND .
+    #ln -s ${TOOLSPATH}/ParamSpider/paramspider.py ${DIRBINPATH}/paramspider ## Já está sendo instalado
 }
 
 PureDNS(){
@@ -416,9 +436,9 @@ Sub404(){
     cd ${TOOLSPATH} 
     git clone https://github.com/r3curs1v3-pr0xy/sub404.git
     cd sub404
-    pip3 install -r requirements.txt
+    $PIPCOMMAND -r requirements.txt
     chmod +x sub404.py
-    ln -s $TOOLSPATH/sub404/sub404.py /usr/local/bin/sub404
+    ln -s ${TOOLSPATH}/sub404/sub404.py ${DIRBINPATH}/sub404
 }
 
 Subfinder(){
@@ -433,16 +453,12 @@ Subjs(){
 
 Telegram-Send(){
     echo -e "${RED}[+]${FUNCNAME[0]}${NC}"
-    pip3 install telegram-send
+    $PIPCOMMAND telegram-send
 }
 
 TurboSearch(){
     echo -e "${RED}[+]${FUNCNAME[0]}${NC}"
-    cd ${TOOLSPATH}
-    git clone https://github.com/helviojunior/turbosearch.git
-    cd turbosearch
-    chmod +x turbosearch.py
-    ln -s ${TOOLSPATH}/turbosearch/turbosearch.py /usr/local/bin/turbosearch
+    $PIPCOMMAND --upgrade turbosearch
 }
 
 Qsreplace(){
@@ -457,7 +473,7 @@ Unfurl(){
 
 Uro(){
     echo -e "${RED}[+]${FUNCNAME[0]}${NC}"
-    pip3 install uro
+    $PIPCOMMAND uro
 }
 
 Waybackurls(){
@@ -475,19 +491,23 @@ WPScan(){
 #INICIO DE CONFIGURACOES FINAIS #################################
 
 PosInstalacao(){
-    if [ $USER == 'root' ]; then
-        mv /root/go/bin/* /usr/local/bin
+    if [ -d "/root/go/bin" ]; then
+        echo -e "${RED}[+]Moving files from /root/go/bin to ${DIRBINPATH}${NC}"
+        mv /root/go/bin/* ${DIRBINPATH}
+    elif [ -d "/go/bin" ]; then
+        echo -e "${RED}[+]Moving files from /go/bin to ${DIRBINPATH}${NC}"
+        mv /go/bin/* ${DIRBINPATH}
+    elif [ -d "/home/$SUDO_USER/go/bin" ]; then
+        echo -e "${RED}[+]Moving files from /home/$SUDO_USER/go/bin to ${DIRBINPATH}${NC}"
+        mv /home/$SUDO_USER/go/bin/* ${DIRBINPATH}
     else
-        mv /home/$SUDO_USER/go/bin/* /usr/local/bin
+        echo -e "${RED}[+]Moving go binary not executed. Nothing to do.${NC}"
     fi
     echo -e "${GREEN}[+] DONE${NC}"
 
 }
 
 #FIM DE CONFIGURACOES FINAIS #################################
-
-
-
 
 callRequirements(){
     setupEnvironment
@@ -496,6 +516,8 @@ callRequirements(){
 }
 
 callInstallTools(){
+    Altdns
+    Alterx
     Amass
     Anew
     Antiburl
@@ -538,7 +560,6 @@ callInstallTools(){
     ParamSpider
     ShufleDNS
     Sub404
-    Subdomains
     Subfinder
     Subjs
     Telegram-Send
